+++ android/app/src/main/java/com/example/omkar/android/fragments/StudentDetailsFragment.java
@@ -83,7 +83,7 @@ public class StudentDetailsFragment extends Fragment{
             Log.d("DATE COUNT", String.valueOf(dateCount));
             Log.d("DAY COUNT", String.valueOf(dayCount));
             // add info
-            mStudentDetailsList.add(new String[]{String.valueOf(i + 1), String.valueOf(c.getInt(1)), String.valueOf(c.getInt(2)), dayCount != 0?String.valueOf((float)((dateCount * 100) / dayCount)):"0"});
+            mStudentDetailsList.add(new String[]{String.valueOf(i + 1), String.valueOf(c.getFloat(1)), String.valueOf(c.getFloat(2)), dayCount != 0?String.valueOf((float)((dateCount * 100) / dayCount)):"0"});
             if(!c.moveToNext()){break;}
         }
         c.close();
@@ -111,7 +111,7 @@ public class StudentDetailsFragment extends Fragment{
         Collections.sort(mStudentDetailsList, new Comparator<String[]>() {
             @Override
             public int compare(String[] s1, String[] s2) {
-                return s1[2].compareTo(s2[2]);
+                return Float.valueOf(s1[2]).compareTo(Float.valueOf(s2[2]));
             }
         });
         mStudentDetailsAdapter.notifyDataSetChanged();
